Some refactoring on Party Editor Implemented a new lighter Browser utility class Some initial stubs for Gambits
