+++ ToyBox/classes/MainUI/PartyEditor/BrainEditor.cs
@@ -1,0 +1,35 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using UnityEngine;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using Kingmaker;
+using Kingmaker.Blueprints.Classes;
+using Kingmaker.Blueprints.Classes.Spells;
+using Kingmaker.Designers;
+using Kingmaker.EntitySystem.Entities;
+using Kingmaker.EntitySystem.Stats;
+using Kingmaker.UnitLogic;
+using ToyBox.Multiclass;
+using Alignment = Kingmaker.Enums.Alignment;
+using ModKit;
+using static ModKit.UI;
+using ModKit.Utility;
+using ToyBox.classes.Infrastructure;
+using Kingmaker.PubSubSystem;
+using Kingmaker.Blueprints;
+using Kingmaker.UnitLogic.FactLogic;
+using Kingmaker.UnitLogic.Parts;
+using static Kingmaker.Utility.UnitDescription.UnitDescription;
+
+
+namespace ToyBox {
+    public partial class PartyEditor {
+        public static void OnBrainGUI(UnitEntityData ch) {
+        }
+    }
+}
