+++ ToyBox/classes/MainUI/PartyEditor/PartyEditor.cs
@@ -1,0 +1,344 @@
+﻿// Copyright < 2021 > Narria (github user Cabarius) - License: MIT
+using UnityEngine;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using Kingmaker;
+using Kingmaker.Blueprints.Classes;
+using Kingmaker.Blueprints.Classes.Spells;
+using Kingmaker.Designers;
+using Kingmaker.EntitySystem.Entities;
+using Kingmaker.EntitySystem.Stats;
+using Kingmaker.UnitLogic;
+using ToyBox.Multiclass;
+using Alignment = Kingmaker.Enums.Alignment;
+using ModKit;
+using static ModKit.UI;
+using ModKit.Utility;
+using ToyBox.classes.Infrastructure;
+using Kingmaker.PubSubSystem;
+using Kingmaker.Blueprints;
+using Kingmaker.UnitLogic.FactLogic;
+using Kingmaker.UnitLogic.Parts;
+
+namespace ToyBox {
+    public partial class PartyEditor {
+        public static Settings settings => Main.settings;
+
+        private enum ToggleChoice {
+            Classes,
+            Stats,
+            Facts,
+            Features,
+            Buffs,
+            Abilities,
+            Spells,
+            Brain,
+            None,
+        };
+
+        private static ToggleChoice selectedToggle = ToggleChoice.None;
+        private static int selectedCharacterIndex = 0;
+        private static UnitEntityData charToAdd = null;
+        private static UnitEntityData charToRecruit = null;
+        private static UnitEntityData charToRemove = null;
+        private static UnitEntityData charToUnrecruit = null;
+        private static bool editMultiClass = false;
+        private static UnitEntityData multiclassEditCharacter = null;
+        private static int respecableCount = 0;
+        private static int recruitableCount = 0;
+        private static int selectedSpellbook = 0;
+        private static (string, string) nameEditState = (null, null);
+        public static int selectedSpellbookLevel = 0;
+        private static bool editSpellbooks = false;
+        private static UnitEntityData spellbookEditCharacter = null;
+        private static readonly Dictionary<string, int> statEditorStorage = new();
+        public static Dictionary<string, Spellbook> SelectedSpellbook = new();
+
+        public static List<UnitEntityData> GetCharacterList() {
+            var partyFilterChoices = CharacterPicker.GetPartyFilterChoices();
+            if (partyFilterChoices == null) { return null; }
+            return partyFilterChoices[Main.settings.selectedPartyFilter].func();
+        }
+
+        private static UnitEntityData GetSelectedCharacter() {
+            var characterList = GetCharacterList();
+            if (characterList == null || characterList.Count == 0) return null;
+            if (selectedCharacterIndex >= characterList.Count) selectedCharacterIndex = 0;
+            return characterList[selectedCharacterIndex];
+        }
+        public static void ResetGUI() {
+            selectedCharacterIndex = 0;
+            selectedSpellbook = 0;
+            selectedSpellbookLevel = 0;
+            CharacterPicker.partyFilterChoices = null;
+            Main.settings.selectedPartyFilter = 0;
+        }
+
+        // This bit of kludge is added in order to tell whether our generic actions are being accessed from this screen or the Search n' Pick
+        public static bool IsOnPartyEditor() => Main.settings.selectedTab == 2;
+
+        public static void ActionsGUI(UnitEntityData ch) {
+            var player = Game.Instance.Player;
+            Space(25);
+            if (!player.PartyAndPets.Contains(ch) && player.AllCharacters.Contains(ch)) {
+                ActionButton("Add", () => { charToAdd = ch; }, Width(150));
+                Space(25);
+            }
+            else if (player.ActiveCompanions.Contains(ch)) {
+                ActionButton("Remove", () => { charToRemove = ch; }, Width(150));
+                Space(25);
+            }
+            else if (!player.AllCharacters.Contains(ch)) {
+                recruitableCount++;
+                ActionButton("Recruit".cyan(), () => { charToRecruit = ch; }, Width(150));
+                Space(25);
+            }
+            if (player.AllCharacters.Contains(ch) && !ch.IsStoryCompanion()) {
+                ActionButton("Unrecruit".cyan(), () => { charToUnrecruit = ch; charToRemove = ch; }, Width(150));
+                Space(25);
+
+            }
+            else {
+                Space(178);
+            }
+            if (RespecHelper.GetRespecableUnits().Contains(ch)) {
+                respecableCount++;
+                ActionButton("Respec".cyan(), () => { Actions.ToggleModWindow(); RespecHelper.Respec(ch); }, Width(150));
+            }
+            else {
+                Space(153);
+            }
+#if false
+            Space(25);
+            ActionButton("Log Caster Info", () => CasterHelpers.GetOriginalCasterLevel(ch.Descriptor),
+                AutoWidth());
+#endif
+        }
+        public static void OnGUI() {
+            var player = Game.Instance.Player;
+            var filterChoices = CharacterPicker.GetPartyFilterChoices();
+            if (filterChoices == null) { return; }
+
+            charToAdd = null;
+            charToRecruit = null;
+            charToRemove = null;
+            charToUnrecruit = null;
+            var characterListFunc = TypePicker(
+                null,
+                ref Main.settings.selectedPartyFilter,
+                filterChoices
+                );
+            var characterList = characterListFunc.func();
+            var mainChar = GameHelper.GetPlayerCharacter();
+            if (characterListFunc.name == "Nearby") {
+                Slider("Nearby Distance", ref CharacterPicker.nearbyRange, 1f, 200, 25, 0, " meters", Width(250));
+                characterList = characterList.OrderBy((ch) => ch.DistanceTo(mainChar)).ToList();
+            }
+            Space(20);
+            var chIndex = 0;
+            recruitableCount = 0;
+            respecableCount = 0;
+            var selectedCharacter = GetSelectedCharacter();
+            var isWide = IsWide;
+            if (Main.IsInGame) {
+                using (HorizontalScope()) {
+                    Label($"Party Level ".cyan() + $"{Game.Instance.Player.PartyLevel}".orange().bold(), AutoWidth());
+                    Space(25);
+#if false   // disabled until we fix performance
+                    var encounterCR = CheatsCombat.GetEncounterCr();
+                    if (encounterCR > 0) {
+                        UI.Label($"Encounter CR ".cyan() + $"{encounterCR}".orange().bold(), UI.AutoWidth());
+                    }
+#endif
+                }
+            }
+            List<Action> todo = new();
+            foreach (var ch in characterList) {
+                var classData = ch.Progression.Classes;
+                // TODO - understand the difference between ch.Progression and ch.Descriptor.Progression
+                var progression = ch.Descriptor.Progression;
+                var xpTable = progression.ExperienceTable;
+                var level = progression.CharacterLevel;
+                var mythicLevel = progression.MythicLevel;
+                var spellbooks = ch.Spellbooks;
+                var spellCount = spellbooks.Sum((sb) => sb.GetAllKnownSpells().Count());
+                var isOnTeam = player.AllCharacters.Contains(ch);
+                using (HorizontalScope()) {
+                    var name = ch.CharacterName;
+                    if (Game.Instance.Player.AllCharacters.Contains(ch)) {
+                        if (isWide) {
+                            if (EditableLabel(ref name, ref nameEditState, 200, n => n.orange().bold(), MinWidth(100), MaxWidth(600))) {
+                                ch.Descriptor.CustomName = name;
+                                Main.SetNeedsResetGameUI();
+                            }
+                        }
+                        else
+                            if (EditableLabel(ref name, ref nameEditState, 200, n => n.orange().bold(), Width(230))) {
+                            ch.Descriptor.CustomName = name;
+                            Main.SetNeedsResetGameUI();
+                        }
+                    }
+                    else {
+                        if (isWide)
+                            Label(ch.CharacterName.orange().bold(), MinWidth(100), MaxWidth(600));
+                        else
+                            Label(ch.CharacterName.orange().bold(), Width(230));
+                    }
+                    Space(5);
+                    var distance = mainChar.DistanceTo(ch); ;
+                    Label(distance < 1 ? "" : distance.ToString("0") + "m", Width(75));
+                    Space(5);
+                    int nextLevel;
+                    for (nextLevel = level; progression.Experience >= xpTable.GetBonus(nextLevel + 1) && xpTable.HasBonusForLevel(nextLevel + 1); nextLevel++) { }
+                    if (nextLevel <= level || !isOnTeam)
+                        Label((level < 10 ? "   lvl" : "   lv").green() + $" {level}", Width(90));
+                    else
+                        Label((level < 10 ? "  " : "") + $"{level} > " + $"{nextLevel}".cyan(), Width(90));
+                    // Level up code adapted from Bag of Tricks https://www.nexusmods.com/pathfinderkingmaker/mods/2
+                    if (player.AllCharacters.Contains(ch)) {
+                        if (xpTable.HasBonusForLevel(nextLevel + 1)) {
+                            ActionButton("+1", () => {
+                                progression.AdvanceExperienceTo(xpTable.GetBonus(nextLevel + 1), true);
+                            }, Width(63));
+                        }
+                        else { Label("max", Width(63)); }
+                    }
+                    else { Space(66); }
+                    Space(10);
+                    var nextML = progression.MythicExperience;
+                    if (nextML <= mythicLevel || !isOnTeam)
+                        Label((mythicLevel < 10 ? "  my" : "  my").green() + $" {mythicLevel}", Width(90));
+                    else
+                        Label((level < 10 ? "  " : "") + $"{mythicLevel} > " + $"{nextML}".cyan(), Width(90));
+                    if (player.AllCharacters.Contains(ch)) {
+                        if (progression.MythicExperience < 10) {
+                            ActionButton("+1", () => {
+                                progression.AdvanceMythicExperience(progression.MythicExperience + 1, true);
+                            }, Width(63));
+                        }
+                        else { Label("max", Width(63)); }
+                    }
+                    else { Space(66); }
+                    Space(30);
+                    if (!isWide) ActionsGUI(ch);
+                    Wrap(!IsWide, 283, 0);
+                    var prevSelectedChar = selectedCharacter;
+                    var showClasses = ch == selectedCharacter && selectedToggle == ToggleChoice.Classes;
+                    if (DisclosureToggle($"{classData.Count} Classes", ref showClasses)) {
+                        if (showClasses) {
+                            selectedCharacter = ch; selectedToggle = ToggleChoice.Classes; Mod.Trace($"selected {ch.CharacterName}");
+                        }
+                        else { selectedToggle = ToggleChoice.None; }
+                    }
+                    var showStats = ch == selectedCharacter && selectedToggle == ToggleChoice.Stats;
+                    if (DisclosureToggle("Stats", ref showStats, 125)) {
+                        if (showStats) { selectedCharacter = ch; selectedToggle = ToggleChoice.Stats; }
+                        else { selectedToggle = ToggleChoice.None; }
+                    }
+                    Wrap(IsNarrow, 279);
+                    //var showFacts = ch == selectedCharacter && selectedToggle == ToggleChoice.Facts;
+                    //if (UI.DisclosureToggle("Facts", ref showFacts, 125)) {
+                    //    if (showFacts) { selectedCharacter = ch; selectedToggle = ToggleChoice.Facts; }
+                    //    else { selectedToggle = ToggleChoice.None; }
+                    //}
+                    var showFeatures = ch == selectedCharacter && selectedToggle == ToggleChoice.Features;
+                    if (DisclosureToggle("Features", ref showFeatures, 150)) {
+                        if (showFeatures) { selectedCharacter = ch; selectedToggle = ToggleChoice.Features; }
+                        else { selectedToggle = ToggleChoice.None; }
+                    }
+                    var showBuffs = ch == selectedCharacter && selectedToggle == ToggleChoice.Buffs;
+                    if (DisclosureToggle("Buffs", ref showBuffs, 125)) {
+                        if (showBuffs) { selectedCharacter = ch; selectedToggle = ToggleChoice.Buffs; }
+                        else { selectedToggle = ToggleChoice.None; }
+                    }
+                    Wrap(IsNarrow, 304);
+                    var showAbilities = ch == selectedCharacter && selectedToggle == ToggleChoice.Abilities;
+                    if (DisclosureToggle("Abilities", ref showAbilities, 125)) {
+                        if (showAbilities) { selectedCharacter = ch; selectedToggle = ToggleChoice.Abilities; }
+                        else { selectedToggle = ToggleChoice.None; }
+                    }
+                    Space(10);
+                    if (spellbooks.Count() > 0) {
+                        var showSpells = ch == selectedCharacter && selectedToggle == ToggleChoice.Spells;
+                        if (DisclosureToggle($"{spellCount} Spells", ref showSpells)) {
+                            if (showSpells) { selectedCharacter = ch; selectedToggle = ToggleChoice.Spells; }
+                            else { selectedToggle = ToggleChoice.None; }
+                        }
+                    }
+                    else { Space(180); }
+                    var showBrain = ch == selectedCharacter && selectedToggle == ToggleChoice.Brain;
+#if false
+                    if (DisclosureToggle("Gambits", ref showBrain, 125)) {
+                        if (showBrain) { selectedCharacter = ch; selectedToggle = ToggleChoice.Brain; }
+                        else { selectedToggle = ToggleChoice.None; }
+                    }
+                    Space(10);
+#endif
+                    if (isWide) ActionsGUI(ch);
+                    if (prevSelectedChar != selectedCharacter) {
+                        selectedSpellbook = 0;
+                    }
+                }
+                //if (!UI.IsWide && (selectedToggle != ToggleChoice.Stats || ch != selectedCharacter)) {
+                //    UI.Div(20, 20);
+                //}
+                if (selectedCharacter != spellbookEditCharacter) {
+                    editSpellbooks = false;
+                    spellbookEditCharacter = null;
+                }
+                if (selectedCharacter != multiclassEditCharacter) {
+                    editMultiClass = false;
+                    multiclassEditCharacter = null;
+                }
+                if (ch == selectedCharacter && selectedToggle == ToggleChoice.Classes) {
+                    OnClassesGUI(ch, classData, selectedCharacter);
+                }
+                if (ch == selectedCharacter && selectedToggle == ToggleChoice.Stats) {
+                    OnStatsGUI(ch);
+                }
+                //if (ch == selectedCharacter && selectedToggle == ToggleChoice.Facts) {
+                //    todo = FactsEditor.OnGUI(ch, ch.Facts.m_Facts);
+                //}
+                if (ch == selectedCharacter && selectedToggle == ToggleChoice.Features) {
+                    todo = FactsEditor.OnGUI(ch, ch.Progression.Features.Enumerable.ToList());
+                }
+                if (ch == selectedCharacter && selectedToggle == ToggleChoice.Buffs) {
+                    todo = FactsEditor.OnGUI(ch, ch.Descriptor.Buffs.Enumerable.ToList());
+                }
+                if (ch == selectedCharacter && selectedToggle == ToggleChoice.Abilities) {
+                    todo = FactsEditor.OnGUI(ch, ch.Descriptor.Abilities.Enumerable.ToList());
+                }
+                if (ch == selectedCharacter && selectedToggle == ToggleChoice.Spells) {
+                    todo = OnSpellsGUI(ch, spellbooks);
+                }
+                if (ch == selectedCharacter && selectedToggle == ToggleChoice.Brain) {
+                    OnBrainGUI(ch);
+                }
+                if (selectedCharacter != GetSelectedCharacter()) {
+                    selectedCharacterIndex = characterList.IndexOf(selectedCharacter);
+                }
+                chIndex += 1;
+            }
+            Space(25);
+            if (recruitableCount > 0) {
+                Label($"{recruitableCount} character(s) can be ".orange().bold() + " Recruited".cyan() + ". This allows you to add non party NPCs to your party as if they were mercenaries".green());
+            }
+            if (respecableCount > 0) {
+                Label($"{respecableCount} character(s)  can be ".orange().bold() + "Respecced".cyan() + ". Pressing Respec will close the mod window and take you to character level up".green());
+                Label("WARNING".yellow().bold() + " The Respec UI is ".orange() + "Non Interruptable".yellow().bold() + " please save before using".orange());
+            }
+            if (recruitableCount > 0 || respecableCount > 0) {
+                Label("WARNING".yellow().bold() + " these features are ".orange() + "EXPERIMENTAL".yellow().bold() + " and uses unreleased and likely buggy code.".orange());
+                Label("BACK UP".yellow().bold() + " before playing with this feature.You will lose your mythic ranks but you can restore them in this Party Editor.".orange());
+            }
+            Space(25);
+            foreach (var action in todo)
+                action();
+            if (charToAdd != null) { UnitEntityDataUtils.AddCompanion(charToAdd); }
+            if (charToRecruit != null) { UnitEntityDataUtils.RecruitCompanion(charToRecruit); }
+            if (charToRemove != null) { UnitEntityDataUtils.RemoveCompanion(charToRemove); }
+            if (charToUnrecruit != null) { charToUnrecruit.Ensure<UnitPartCompanion>().SetState(CompanionState.None); charToUnrecruit.Remove<UnitPartCompanion>(); }
+        }
+    }
+}
